+++ extensions/projects/src/types.ts
@@ -21,6 +21,7 @@ export interface Preferences {
   repoScanPath: string;
   repoScanDepth?: number;
   openNodeWith: OpenWith;
+  openXcodeWith: OpenWith;
   openMavenWith: OpenWith;
   openGradleWith: OpenWith;
   openWith1: OpenWith;
@@ -32,6 +33,7 @@ export enum ProjectType {
   NODE = "node",
   MAVEN = "maven",
   GRADLE = "gradle",
+  XCODE = "xcode",
   UNKNOWN = "unknown",
 }